Create a new locup system for edges
This goes to all four edge, one a time in a ccw direction.
It resets the culture diam and radius, sets the center
as cx,cy and the x0,y0.

It's more accurate than before but the pattern positions
are no longer valid because the diam has changed.  So
this is not the right solution.

The way forward is to not go so close to the edge, not
reset the diam and radius.  Instead, use the averages
of the edges to set the center and set x0,y0 from that.
Or rather, re-create the pattern using the center as
the origin.