Change the timeout on the spo serial.Serial
It's in m99_sim_serial.py.
Changed it from timeout=2 to timeout=1.
It was taking a bit long and I would like to try it like this
to see if I can get it to work.  Tested it a bit with the
locup system and it appeared to work ok.