Create the arec system.
It remembers areas that are set manually, saves them, loads
them, and can move the focus back to areas called by name.

This was done super quick under time pressure because of needing
it for today's experiment with lung transplant explants.

Much still needs to be done to make it easier to use.  But basic
functions work.