Change the arec aut-naming system
Now it starts with an auto-name in this format:
  a###
Each # is a digit.  It starts with these:
  a001
  a002
  ...

The 'a' is the prefix.  It can be changed by doing something
like this:
>>> arec.prefix('b')
In this case, it was changed to 'b'.  Longer prefixes should
work fine but this hasn't been tested.

It always adds to the highest currently stored name.  So if
we have these names and the current prefix is 'a',
  a001
  a703
  a005
  b904
  happy_name
  a8000
it will create:  a704
The a8000 is not the correct length string and gets ignored.
The b904 doesn't use the current prefix and gets ignored.