Implement a new !m system for locup pattern messages
This brings the locup system pattern loading in line with the
scripts I use to create the pattern files.  Also, the pattern
files a cleaner this way.

The old pattern data required any message to come on the same line
as the location coordinates, after a colon like this:
  1 614 691 : 6 central FOVs

The new system allows putting a message before the location line
with which it will be printed.  Like this:
  !m : 6 central FOVs
  1 614 691

The old way is still supported and having a message on the same line
as the location data causes that message to override a previous
message using an !m line.