Move the locup pattern files to the config dir
Since it's now possible to name the file in the locup.config, we can
put the files directly in the config dir.  Like that, they don't
have to be copied to the user/ dir to be used.

Also added a README.txt file to the examples dir just so it wouldn't
be empty.  I don't want git to loose that dir.