Add locup pattern 9, MCTD, NS, equal area
Add a locup pattern (09).  It's for large millicell MCTDs.
It is only for the N and S FOVs so super quick experiments.
It set's the FOVs on the equal-area track, as opposed to the
mean radius track.  The equal-area track uses the radius that
gives the same area in the inside and outside of the track.