Change muwp to use internal coordinates
Before this, muwp defined everything in terms of prior stage
coordinates.  This required the prior stage's coordinates to
be set to whatever was required for the particular plate being
used.  This was a problem in case were wanted to use the muwp
system but wanted to keep the prior stage coordinate system
unchange for another use.

Now, we can manually to a particular fiducial (as defined in the
plate data file) and do (using fiducial x1 as an example):
  u>> set fidu x1
This will reset the muwp psx0 and psy0 values using the current
position and the expected fidu x1 position.

In other parts of the code, going to a particular location no
longer goes directly to it's muwp x y.  Instead it is first
converted to prior stage psx psy.
  psx = x + self.psx0
  psy = y + self.psy0