Finish the "reset all edges" funtions
It's only set up for n_ins == 2 and n_ins == 6.
Also, after testing it looked like some insert edges were not
where I expected.  I wonder if vibrations caused the inserts
to move.  I was testing with no insert holder and dry wells.
In the case where there is no insert holder but fluid in the
well (or dish), would the fluid help dampen any effect of vibration
on the insert or would it make it worse because of boyancy
effects?  Not sure.

Leave the "UNDER CONSTRUCTION" warning.