Create "reset multi edges" using config
The "reset multi edges" using config files replaces the hard
coded "reset all edges".  In the new system, a muwp/plate_xxx.data
file now has info about inserts (not necessarily all) for which
the edges need to be reset and which FOV (N S E or W) to start
from for that insert.  The order of inserts determines the
order used to reset edges.

The data is checked at the time it is loaded from the plate_xxx.data
file.

Resetting data has been added to every plate file in config/muwp/.

The user interface has been changed.
Instead of
  muwp>> reset all edges
it uses
  muwp>> reset multi edges
with this abbreviation also available:
  muwp>> rme