Enforce automatic backup during saving
The backup file will only be created if the file being saved
is 'user/arec.data' and that file already exists.

The backup file user/arec_yyyymmdd_hhmmss.data will have the
timestamp in its filename match exactly the timestamp in
the new user/arec.data file's save time comment line.

Note that loading still doesn't result in an auto save (ie backup)
of the data in ram.  So using 'set' and then loading causes the
loss of that data.