Remove culture diameters from locup pattern files
The culture diameter in the config file is over-written
by the one in the pattern file.  We usually set the
diameter in the config file.  But that won't have any
effect if there is also a diameter set in the pattern
file.  So let's keep them out of the pattern files.